requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
anthropic==0.7.7
python-dotenv==1.0.0
httpx==0.25.2

# PII Detection & Scrubbing
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
